Move BasePage to common
