+++ src/public/common/BasePage.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Helmet } from 'react-helmet';
 import { Page, NavBar, Footer } from 'ot-ui';
 
-import { externalLinks } from '../constants';
+import { externalLinks } from '../../constants';
 
 const BasePage = ({ children }) => (
   <Page
