+++ src/public/disease/Page.js
@@ -6,7 +6,7 @@ import { Helmet } from 'react-helmet';
 
 import { Tabs, Tab } from 'ot-ui';
 
-import BasePage from '../../pages/BasePage';
+import BasePage from '../common/BasePage';
 import Associations from './Associations';
 import Header from './Header';
 import DiseaseOverviewTab from './DiseaseOverviewTab';
